% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_brick.R
\name{brick_ls_works}
\alias{brick_ls_works}
\title{Check that biobricks::brick_ls_remote() can be run
on brick}
\usage{
brick_ls_works(brickname)
}
\arguments{
\item{brickname}{the name of the brick check}
}
\value{
list(valid = boolean, error_msg = string)
}
\description{
Check that biobricks::brick_ls_remote() can be run
on brick
}
