% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_brick.R
\name{file_exists_check}
\alias{file_exists_check}
\title{Check that a file exists and is nonempty}
\usage{
file_exists_check(filename)
}
\arguments{
\item{filename}{to check repo}
}
\value{
list(valid = boolean, error_msg = string)
}
\description{
Check that a file exists and is nonempty
}
