% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_brick.R
\name{valid}
\alias{valid}
\title{Check whether or not the brick is valid}
\usage{
valid(validate_result)
}
\arguments{
\item{validate_result}{a list of validation results returned by validate}
}
\value{
TRUE if the brick passes all tests and FALSE if it does not.
}
\description{
This function checks whether or not a brick contains all of the expected
components for biobricks–ai.
In order to do this, this function tests for the presence of the following
files: dvc.yaml, dvc.lock, .dvc/config, and README.md.
}
