% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_brick.R
\name{validate}
\alias{validate}
\title{Validate the brick by running tests}
\usage{
validate(dir)
}
\arguments{
\item{dir}{of the brick repo}
}
\value{
a list of test results
}
\description{
Validate the brick by running tests
}
