% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_brick.R
\name{errors}
\alias{errors}
\title{Return errors associated with a brick}
\usage{
errors(validate_result)
}
\arguments{
\item{validate_result}{a list of validation results returned by validate}
}
\value{
list of errors
}
\description{
Return errors associated with a brick
}
